Generates css docs when only css directory given
Configurable with styleguide.html which contains header and end tags.
For some reason I cannot get the block to work in helper function,
so the example_html must be given as a parameter for now.

I try to fix that in the future, but seems so strange really.

Check readme for usage instructions